+++ src/widgets/splits/InputCompletionPopup.hpp
@@ -14,16 +14,17 @@ class InputCompletionPopup : public BasePopup
 {
     using ActionCallback = std::function<void(const QString &)>;
 
-    constexpr static int maxEntryCount = 200;
+    constexpr static int MAX_ENTRY_COUNT = 200;
 
 public:
     InputCompletionPopup(QWidget *parent = nullptr);
 
     void updateEmotes(const QString &text, ChannelPtr channel);
     void updateUsers(const QString &text, ChannelPtr channel);
-    virtual bool eventFilter(QObject *, QEvent *event) override;
 
     void setInputAction(ActionCallback callback);
+
+    bool eventFilter(QObject *watched, QEvent *event) override;
 
 protected:
     void showEvent(QShowEvent *event) override;